document processor/requirements.txt
google-generativeai
python-dotenv
watchdog